// src/app/models/game.ts
export interface Game {
    home_team_score: number;
    visitor_team_score: number;
    visitor_team_abbreviation: string;
    home_team_abbreviation: string;
}


/*
Ich brauche in angular ein service, die diese Url aufruft und die Informations mit subscribe in 
einem Component wiedergibt, das sind unsere model: das sind die resulat von diesem url, und hier sind die ergebnis von diesem: 

*/